refactor: 기능 역할 분리
