+++ src/main/java/racingcar/model/Car.java
@@ -11,6 +11,11 @@ public class Car {
         this.name = name;
     }
 
+    public Car(String name, int position) {
+        this.name = name;
+        this.position = position;
+    }
+
     private static boolean canMove() {
         return Randoms.pickNumberInRange(0,9) >= 4;
     }
