+++ src/main/java/racingcar/model/Cars.java
@@ -2,6 +2,7 @@ package racingcar.model;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
 
 public class Cars {
     private final List<Car> carList;
@@ -11,5 +12,20 @@ public class Cars {
     }
     public List<Car> getCarList() {
         return carList;
+    }
+    public List<Car> getWinners() {
+        int maxPosition = carList.stream()
+                .mapToInt(Car::getPosition)
+                .max()
+                .orElse(0);
+
+        return carList.stream()
+                .filter(car -> car.getPosition() == maxPosition)
+                .collect(Collectors.toList());
+    }
+    public void moveCars() {
+        for (Car car : this.carList) {
+            car.move();
+        }
     }
 }
